keep OffHeapVectorValues and all its subclasses in one place
